# Faster inverted index with minheap traversal