… lib/sparse/src/index/search_context.rs