This PR builds on top of https://github.com/qdrant/qdrant/pull/3347 to go even faster.

The key insight is that it is unnecessary to constantly probe all posting list for finding the ones that should be explored.

By leveraging the previously introduced min-heap to drive the search, we can access directly the right positing lists.

### Performance

This brings a a large performance improvement to the criterion benchmarks.
Those are using 50k indexed sparse vectors with around 300 dimensions with a max dimension of 30k.

```
sparse-vector-search-group/mmap-inverted-index-search
                        time:   [8.7720 ms 8.7792 ms 8.7891 ms]
                        change: [-72.942% -72.337% -71.780%] (p = 0.00 < 0.05)
                        Performance has improved.

sparse-vector-search-group/inverted-index-search
                        time:   [10.112 ms 10.142 ms 10.174 ms]
                        change: [-66.641% -66.489% -66.342%] (p = 0.00 < 0.05)
                        Performance has improved.
```

### Notes

1. Given that the posting list cannot be re-ordered anymore, the logic to decide which posting to prune has been changed to target any postings which has the lowest unique min record id. But the pruning logic as not changed.

2. The score seems to not be completely equivalent due to the change of order for float multiplication.

3. I have verified that the python client congruence tests are still green with this change.