@timvisee 2024-01-09
> A consensus [test](https://github.com/qdrant/qdrant/actions/runs/7458986467/job/20296989912?pr=3359) is failing for sparse vector search.
> 
> ```
>   -  [{'id': 4, 'payload': None, 'score': 1.3620001, 'vector': None, 'version': 0},
>   ?                                            ----
>   +  [{'id': 4, 'payload': None, 'score': 1.362, 'vector': None, 'version': 0},
> ```
> 
> My guess is that the minheap does not break ties deterministically which creates slight difference in float precision due to a different multiplication ordering.

Potentially stupid question, but are you sure you're not mutating any other items in the min heap? Asking because that's unsupported.