@agourlay 2024-01-10
@timvisee I fixed the test with https://github.com/qdrant/qdrant/pull/3359/commits/1f6714027faa6c666141518dbb6229cba4262109

I assume the internal ids can be different in replicas leading to a different order of float computation.