@agourlay 2024-01-15
Converted back into a draft as I need to re-evaluate this approach using a realistic dataset.
The data distribution has a strong impact on the performance.