@agourlay 2024-01-17
closing as the approach does not seem to shine with highly skewed data